refactor(agent-tools): extract makeAgentMgmtTools into agent-mgmt-tools.ts
