+++ chatbot/lib/agent-tools.ts
@@ -6,17 +6,10 @@
  */
 
 import { tool } from "@anthropic-ai/claude-agent-sdk";
-import {
-  installAgent,
-  uninstallAgent,
-  loadAgent,
-  unloadAgent,
-  isLoaded,
-  getAgentStatus,
-  getAgentLogs,
-} from "@/lib/launchd";
-import { cancelProcessing } from "@/a2a/lib/processing-registry";
 import type { AgentDef } from "@@/lib/agents";
+import { MGMT_TOOL } from "./agent-mgmt-tools";
+
+export * from "./agent-mgmt-tools";
 import { formatScheduleDisplay } from "@@/lib/agents-config-schemas";
 import { plistLabel } from "@@/lib/plist-generate";
 import {
@@ -149,17 +142,6 @@ export const justificationField = z
   })
   .describe("Required on every delegation call. Fill this out before handing off.");
 
-// ─── Management tool names ─────────────────────────────────────────────────────
-
-export const MGMT_TOOL = {
-  install: "install_agent",
-  uninstall: "uninstall_agent",
-  load: "load_agent",
-  unload: "unload_agent",
-  status: "check_status",
-  logs: "get_logs",
-} as const;
-
 /** Tool name for firing the agent script in the background (start_run_script_* pattern). */
 export const startRunScriptToolName = (manifestKey: string): string => `start_${manifestKey}`;
 /** Tool name for polling a previously started script run (await_run_script_* pattern). */
@@ -189,114 +171,6 @@ export const startEscalateToToolName = (manifestKey: string): string =>
 /** Tool name for collecting the result of a start_escalate_to_* call. */
 export const awaitEscalateToToolName = (manifestKey: string): string =>
   `await_escalate_to_${manifestKey}`;
-
-// ─── Management tools factory ─────────────────────────────────────────────────
-
-/** Returns the 6 per-agent launchd management tools for use in an inner MCP server. */
-export function makeAgentMgmtTools(agent: AgentDef) {
-  const installTool = tool(
-    MGMT_TOOL.install,
-    `Build and install only the ${agent.displayName} agent (scoped tsup build → deploy script → write plist → bootstrap)`,
-    {},
-    async () => {
-      const { loaded, skipped } = await installAgent(agent);
-      if (skipped) {
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: `✅ ${agent.displayName} is not scheduling-enabled — launchd install skipped.`,
-            },
-          ],
-        };
-      }
-      return {
-        content: [
-          {
-            type: "text" as const,
-            text: loaded
-              ? `✅ ${agent.displayName} installed and loaded.`
-              : `⚠️ ${agent.displayName} plist written but not loaded — check launchctl.`,
-          },
-        ],
-      };
-    },
-  );
-
-  const uninstallTool = tool(
-    MGMT_TOOL.uninstall,
-    `Unload and delete only the ${agent.displayName} agent plist`,
-    {},
-    async () => {
-      cancelProcessing(agent.manifestKey);
-      await uninstallAgent(agent);
-      return {
-        content: [
-          { type: "text" as const, text: `✅ ${agent.displayName} unloaded and plist deleted.` },
-        ],
-      };
-    },
-  );
-
-  const loadTool = tool(
-    MGMT_TOOL.load,
-    `Bootstrap (load) the ${agent.displayName} plist into launchd`,
-    {},
-    async () => {
-      await loadAgent(agent);
-      const loaded = await isLoaded(plistLabel(agent));
-      return {
-        content: [
-          {
-            type: "text" as const,
-            text: loaded
-              ? `✅ ${agent.displayName} loaded.`
-              : `⚠️ ${agent.displayName} bootstrap attempted but not showing as loaded.`,
-          },
-        ],
-      };
-    },
-  );
-
-  const unloadTool = tool(
-    MGMT_TOOL.unload,
-    `Bootout (unload) the ${agent.displayName} from launchd`,
-    {},
-    async () => {
-      cancelProcessing(agent.manifestKey);
-      await unloadAgent(agent);
-      return {
-        content: [{ type: "text" as const, text: `✅ ${agent.displayName} unloaded.` }],
-      };
-    },
-  );
-
-  const checkStatusTool = tool(
-    MGMT_TOOL.status,
-    `Get launchd state, PID, last exit code, and loaded status for ${agent.displayName}`,
-    {},
-    async () => {
-      const [{ state, pid, lastExitCode, raw }, loaded] = await Promise.all([
-        getAgentStatus(agent),
-        isLoaded(plistLabel(agent)),
-      ]);
-      const summary = `loaded=${loaded}  state=${state ?? "unknown"}  pid=${pid ?? "-"}  last_exit=${lastExitCode ?? "-"}`;
-      return { content: [{ type: "text" as const, text: `${summary}\n\n${raw}` }] };
-    },
-  );
-
-  const getLogsTool = tool(
-    MGMT_TOOL.logs,
-    `Read recent log output for ${agent.displayName}`,
-    { lines: z.number().optional().describe("Number of lines to return (default 100)") },
-    async ({ lines }) => {
-      const output = await getAgentLogs(agent, lines);
-      return { content: [{ type: "text" as const, text: output }] };
-    },
-  );
-
-  return [installTool, uninstallTool, loadTool, unloadTool, checkStatusTool, getLogsTool];
-}
 
 // ─── Script run tools ─────────────────────────────────────────────────────────
